+++ tiled/client/context.py
@@ -236,19 +236,15 @@ class Context:
             # where the context starts and stops and the wrapped ASGI app.
             # We are abusing it slightly to enable interactive use of the
             # TestClient.
-            if sys.version_info < (3, 9):
-                import atexit
 
-                atexit.register(client.__exit__)
-            else:
-                import threading
+            import threading
 
-                # The threading module has its own (internal) atexit
-                # mechanism that runs at thread shutdown, prior to the atexit
-                # mechanism that runs at interpreter shutdown.
-                # We need to intervene at that layer to close the portal, or else
-                # we will wait forever for a thread run by the portal to join().
-                threading._register_atexit(client.__exit__)
+            # The threading module has its own (internal) atexit
+            # mechanism that runs at thread shutdown, prior to the atexit
+            # mechanism that runs at interpreter shutdown.
+            # We need to intervene at that layer to close the portal, or else
+            # we will wait forever for a thread run by the portal to join().
+            threading._register_atexit(client.__exit__)
 
         self.http_client = client
         self._verify = verify
@@ -385,6 +381,15 @@ class Context:
         """
         uri = httpx.URL(uri)
         node_path_parts = []
+        # Ensure that HTTPS is used if available
+        # Logic will follow only one redirect, it is intended ONLY to toggle HTTPS.
+        # The redirect will be followed only if the netloc host is identical to the original.
+        if uri.scheme == "http":
+            response_from_http = httpx.get(uri)
+            if response_from_http.is_redirect:
+                redirect_uri = httpx.URL(response_from_http.headers["location"])
+                if redirect_uri.scheme == "https" and redirect_uri.host == uri.host:
+                    uri = redirect_uri
         if "/metadata" in uri.path:
             api_path, _, node_path = uri.path.partition("/metadata")
             api_uri = uri.copy_with(path=api_path)